pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[openai]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
]
